+++ CHANGELOG.md
@@ -19,6 +19,11 @@
 * Removed `withShortDebug` utility method. Use `withDebug` instead, which now always logs a single
   line upon completion. This API simplification mirrors a recent change to `hoist-core`.
 
+### ⚙️ Technical
+
+* New `Exception.timeout()` util to throw exceptions explicitly marked as timeouts, used by
+  `Promise.timeout` extension.
+
 [Commit Log](https://github.com/xh/hoist-react/compare/v41.0.0...develop)
 
 ## v41.0.0 - 2021-07-01